Big Updates
Created a dti_fit script to run DTI Fit
Added a bedpostx_batch script to run bedpostx
More flag options in preproc.sh to run each step of the pipeline simultaneously or separately
Note: There was a bedpostx_gpu error on the server - this should be remedied by calling a newer version of FSL in the sbatch script.
NOTE: Will still need to update script with finalized register_batch script (This portion of the script needs to be completed).
DP